WebService complete

+++ BikeStores/BikeDataLibrary/ServiceModels/ServiceProduct.cs
@@ -10,7 +10,7 @@ namespace BikeDataLibrary.ServiceModels
     {
         public int ProductId { get; set; }
         public string ProductName { get; set; }
-        public int ModelYear { get; set; }
+        public short ModelYear { get; set; }
         public decimal ListPrice { get; set; }
         public int BrandId { get; set; }
         public int CategoryId { get; set; }
